Add registration

+++ src/main/java/byAJ/Securex/repositories/BookRepository.java
@@ -4,5 +4,6 @@ import byAJ.Securex.models.Book;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
+
 }
